+++ lib/tty/editor.rb
@@ -155,11 +155,14 @@ module TTY
     #   whether or not show commands menu, true by default
     # @param [Boolean] enable_color
     #   disable or force prompt coloring, defaults to nil
+    # @param [Symbol] menu_interrupt
+    #   how to handle Ctrl+C key interrupt out of :error, :signal, :exit, :noop
     #
     # @api public
     def initialize(command: nil, raise_on_failure: false, show_menu: true,
                    prompt: "Select an editor?", env: {}, enable_color: nil,
-                   input: $stdin, output: $stdout, &block)
+                   input: $stdin, output: $stdout, menu_interrupt: :error,
+                   &block)
       @env = env
       @command = nil
       @input = input
@@ -168,6 +171,7 @@ module TTY
       @enable_color = enable_color
       @show_menu = show_menu
       @prompt = prompt
+      @menu_interrupt = menu_interrupt
 
       block.(self) if block
 
@@ -316,7 +320,8 @@ module TTY
     def choose_exec_from(execs)
       if @show_menu && execs.size > 1
         prompt = TTY::Prompt.new(input: @input, output: @output, env: @env,
-                                 enable_color: @enable_color)
+                                 enable_color: @enable_color,
+                                 interrupt: @menu_interrupt)
         exec = prompt.enum_select(@prompt, execs)
         @output.print(prompt.cursor.up + prompt.cursor.clear_line)
         exec